Refactored several duplicated methods into new FpdfTrait and added _putlinks() method to write imported links
