+++ src/Fpdi.php
@@ -23,131 +23,12 @@ use setasign\Fpdi\PdfParser\Type\PdfNull;
 class Fpdi extends FpdfTpl
 {
     use FpdiTrait;
+    use FpdfTrait;
 
     /**
      * FPDI version
      *
      * @string
      */
-    const VERSION = '2.3.7';
-
-    protected function _enddoc()
-    {
-        parent::_enddoc();
-        $this->cleanUp();
-    }
-
-    /**
-     * Draws an imported page or a template onto the page or another template.
-     *
-     * Give only one of the size parameters (width, height) to calculate the other one automatically in view to the
-     * aspect ratio.
-     *
-     * @param mixed $tpl The template id
-     * @param float|int|array $x The abscissa of upper-left corner. Alternatively you could use an assoc array
-     *                           with the keys "x", "y", "width", "height", "adjustPageSize".
-     * @param float|int $y The ordinate of upper-left corner.
-     * @param float|int|null $width The width.
-     * @param float|int|null $height The height.
-     * @param bool $adjustPageSize
-     * @return array The size
-     * @see Fpdi::getTemplateSize()
-     */
-    public function useTemplate($tpl, $x = 0, $y = 0, $width = null, $height = null, $adjustPageSize = false)
-    {
-        if (isset($this->importedPages[$tpl])) {
-            $size = $this->useImportedPage($tpl, $x, $y, $width, $height, $adjustPageSize);
-            if ($this->currentTemplateId !== null) {
-                $this->templates[$this->currentTemplateId]['resources']['templates']['importedPages'][$tpl] = $tpl;
-            }
-            return $size;
-        }
-
-        return parent::useTemplate($tpl, $x, $y, $width, $height, $adjustPageSize);
-    }
-
-    /**
-     * Get the size of an imported page or template.
-     *
-     * Give only one of the size parameters (width, height) to calculate the other one automatically in view to the
-     * aspect ratio.
-     *
-     * @param mixed $tpl The template id
-     * @param float|int|null $width The width.
-     * @param float|int|null $height The height.
-     * @return array|bool An array with following keys: width, height, 0 (=width), 1 (=height), orientation (L or P)
-     */
-    public function getTemplateSize($tpl, $width = null, $height = null)
-    {
-        $size = parent::getTemplateSize($tpl, $width, $height);
-        if ($size === false) {
-            return $this->getImportedPageSize($tpl, $width, $height);
-        }
-
-        return $size;
-    }
-
-    /**
-     * @inheritdoc
-     * @throws CrossReferenceException
-     * @throws PdfParserException
-     */
-    protected function _putimages()
-    {
-        $this->currentReaderId = null;
-        parent::_putimages();
-
-        foreach ($this->importedPages as $key => $pageData) {
-            $this->_newobj();
-            $this->importedPages[$key]['objectNumber'] = $this->n;
-            $this->currentReaderId = $pageData['readerId'];
-            $this->writePdfType($pageData['stream']);
-            $this->_put('endobj');
-        }
-
-        foreach (\array_keys($this->readers) as $readerId) {
-            $parser = $this->getPdfReader($readerId)->getParser();
-            $this->currentReaderId = $readerId;
-
-            while (($objectNumber = \array_pop($this->objectsToCopy[$readerId])) !== null) {
-                try {
-                    $object = $parser->getIndirectObject($objectNumber);
-                } catch (CrossReferenceException $e) {
-                    if ($e->getCode() === CrossReferenceException::OBJECT_NOT_FOUND) {
-                        $object = PdfIndirectObject::create($objectNumber, 0, new PdfNull());
-                    } else {
-                        throw $e;
-                    }
-                }
-
-                $this->writePdfType($object);
-            }
-        }
-
-        $this->currentReaderId = null;
-    }
-
-    /**
-     * @inheritdoc
-     */
-    protected function _putxobjectdict()
-    {
-        foreach ($this->importedPages as $key => $pageData) {
-            $this->_put('/' . $pageData['id'] . ' ' . $pageData['objectNumber'] . ' 0 R');
-        }
-
-        parent::_putxobjectdict();
-    }
-
-    /**
-     * @inheritdoc
-     */
-    protected function _put($s, $newLine = true)
-    {
-        if ($newLine) {
-            $this->buffer .= $s . "\n";
-        } else {
-            $this->buffer .= $s;
-        }
-    }
+    const VERSION = '2.4.0';
 }
